backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asyncpg==0.31.0
cffi==2.0.0
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
idna==3.11
jwt==1.4.0
Mako==1.3.11
MarkupSafe==3.0.3
pwdlib==0.3.0
pycparser==3.0
pydantic==2.13.2
pydantic-settings==2.13.1
pydantic_core==2.46.2
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
